service - buyer interface

+++ Lab02/ClientSideBuyer/client_side_buyer_server.py
@@ -5,22 +5,26 @@ from flask import Flask, request
 host, port, app = '127.0.0.1', 65534, Flask(__name__)
 
 
-@app.route('/api/v1/buyer/items', methods=['GET'])
-def enlist():
-    items_list = services.get()
+@app.route('/api/v1/buyer/cart', methods=['GET'])
+def display_cart():
+    buyer_id = request.args.get('buyer_id')
+    if not buyer_id:
+        string_util.error['error_message'] = string_util.missing_request_parameters.format(request_parameters='buyer id')
+        return Util.get_response_object(string_util.error, 401)
+    items_list = services.display(buyer_id=buyer_id)
     return Util.get_response_object(items_list, 200)
 
 
 @app.route('/api/v1/buyer/items/search', methods=['POST'])
-def search():
+def search_items():
     items_list = services.search(request.get_json(force=True))
     return Util.get_response_object(items_list, 200)
 
 
-@app.route('/api/v1/buyer/items', methods=['POST'])
-def add():
+@app.route('/api/v1/buyer/cart/items', methods=['POST'])
+def add_items_to_cart():
     payload = request.get_json(force=True)
-    response = services.post(data=payload)
+    response = services.add_item_to_cart(data=payload)
     if response is not None:
         string_util.error['error_message'] = response
         return Util.get_response_object(string_util.error, 401)
@@ -28,10 +32,10 @@ def add():
         return Util.get_response_object(response=None, status_code=200)
 
 
-@app.route('/api/v1/buyer/items', methods=['PUT', 'UPDATE'])
-def update():
+@app.route('/api/v1/buyer/cart/item', methods=['PUT', 'UPDATE'])
+def remove_item_from_cart():
     payload = request.get_json(force=True)
-    response = services.put_or_update(data=payload)
+    response = services.remove_item_from_cart(data=payload)
     if response is not None:
         string_util.error['error_message'] = response
         return Util.get_response_object(string_util.error, 401)
@@ -39,8 +43,12 @@ def update():
         return Util.get_response_object(response=None, status_code=200)
 
 
-@app.route('/api/v1/buyer/items', methods=['DELETE'])
-def delete():
+@app.route('/api/v1/buyer/cart', methods=['DELETE'])
+def clear_cart():
+    buyer_id = request.args.get('buyer_id')
+    if not buyer_id:
+        string_util.error['error_message'] = string_util.missing_request_parameters.format(request_parameters='buyer id')
+        return Util.get_response_object(string_util.error, 401)
     response = services.delete()
     if response is not None:
         string_util.error['error_message'] = response
@@ -51,12 +59,24 @@ def delete():
 
 @app.route('/api/v1/buyer/account/create', methods=['POST'])
 def create_account():
-    return None
+    payload = request.get_json(force=True)
+    response = services.create_buyer_account(data=payload)
+    if response is not None:
+        string_util.error['error_message'] = response
+        return Util.get_response_object(string_util.error, 401)
+    else:
+        return Util.get_response_object(response=None, status_code=200)
 
 
 @app.route('/api/v1/buyer/account/login', methods=['POST'])
 def login():
-    return None
+    payload = request.get_json(force=True)
+    response = services.login_buyer_account(data=payload)
+    if 'Error' in response:
+        string_util.error['error_message'] = response
+        return Util.get_response_object(string_util.error, 401)
+    else:
+        return Util.get_response_object(response=response, status_code=200)
 
 
 @app.route('/api/v1/buyer/account/logout', methods=['GET'])
@@ -73,5 +93,17 @@ def make_purchase():
         return Util.get_response_object(string_util.error, 401)
     else:
         return Util.get_response_object(response=None, status_code=200)
+
+
+@app.route('/api/v1/buyer/feedback', methods=['POST'])
+def feedback():
+    payload = request.get_json(force=True)
+    response = services.submit_feedback(data=payload)
+    if response is not None:
+        string_util.error['error_message'] = response
+        return Util.get_response_object(string_util.error, 401)
+    else:
+        return Util.get_response_object(response=None, status_code=200)
+
 
 app.run(host, port)
